Separate stream-related logic into their own file
